+++ webapp/common/tests/models/test_base_model.py
@@ -1,4 +1,4 @@
-from common.models.base_model import BaseModel
+from common.models import BaseModel
 from django.db import models
 from django.test import TestCase
