+++ webapp/common/tests/views/test_base_viewset.py
@@ -1,4 +1,4 @@
-from common.models.base_model import BaseModel
+from common.models import BaseModel
 from common.views import BaseViewSet
 from django.db import models
 from django.test import TestCase
