+++ webapp/profiles/models/profile.py
@@ -1,6 +1,6 @@
 import uuid
 
-from common.models.base_model_with_soft_delete import BaseModelWithSoftDelete
+from common.models import BaseModelWithSoftDelete
 from django.conf import settings
 from django.db import models
 from django.db.models.signals import m2m_changed
